requirements.txt
filelock==3.15.4
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.4.0
sympy==1.13.1
timm==0.5.4
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
